Change PullRequest behavior so we merge the PR into master
Because GitHub tends to send the webhook event before the synthetic
merge ref has been updated there's no sha that is known before hand.
This can/has lead to builds of old code, the solution was to wait a bit
and retry the build and hope GitHub updated the merge ref.

This was taken from how cloud.drone.io currently does things, which was
likely learned from TravisCI and/or other systems. GitHub does know what
the commit pushed was and does send that in the webhook payload, so we
will do the synthetic merge ourselves.